# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "覇権小説エンジン API",
    "version": "3.0"
  },
  "paths": {
...
    "/api/refine_erotic": {
      "post": {
        "summary": "Refine Erotic",
        "operationId": "refine_erotic_api_refine_erotic_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefineEroticRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/easy_mode/generate": {
      "post": {
        "summary": "Generate Easy",
        "operationId": "generate_easy_api_easy_mode_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EasyModeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plots/plan_generation": {
      "post": {
        "summary": "Plan Generation",
        "operationId": "plan_generation_api_plots_plan_generation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanGenerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/episodes/retry_failed": {
      "post": {
        "summary": "Retry Failed Episodes",
        "operationId": "retry_failed_episodes_api_episodes_retry_failed_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetryFailedRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/episodes/generate": {
      "post": {
        "summary": "Generate Episodes",
        "operationId": "generate_episodes_api_episodes_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EpisodeGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/episodes/generate_candidates": {
      "post": {
        "summary": "Generate Episodes Candidates",
        "operationId": "generate_episodes_candidates_api_episodes_generate_candidates_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EpisodeGenerateCandidatesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plots/expand": {
      "post": {
        "summary": "Expand Plots",
        "operationId": "expand_plots_api_plots_expand_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlotExpandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plots/expand_candidates": {
      "post": {
        "summary": "Expand Plots Candidates",
        "operationId": "expand_plots_candidates_api_plots_expand_candidates_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlotExpandCandidatesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plots/rebuild": {
      "post": {
        "summary": "Rebuild Plots",
        "operationId": "rebuild_plots_api_plots_rebuild_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlotRebuildRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/critique/optimize": {
      "post": {
        "summary": "Critique Optimize",
        "operationId": "critique_optimize_api_critique_optimize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CritiqueOptimizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/plots/audit": {
      "post": {
        "summary": "Audit Plan",
        "operationId": "audit_plan_api_plots_audit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuditPlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chapters/import": {
      "post": {
        "summary": "Import Chapter",
        "operationId": "import_chapter_api_chapters_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChapterImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChapterImportRequest": {
        "properties": {
          "api_key": {
            "type": "string",
            "title": "Api Key"
          },
          "book_id": {
            "type": "integer",
            "title": "Book Id"
          },
          "ep_num": {
            "type": "integer",
            "title": "Ep Num"
          },
          "import_text": {
            "type": "string",
            "title": "Import Text"
          },
          "do_refine": {
            "type": "boolean",
            "title": "Do Refine"
          }
        },
        "type": "object",
        "required": [
          "api_key",
          "book_id",
          "ep_num",
          "import_text",
          "do_refine"
        ],
        "title": "ChapterImportRequest",
        "description": "章インポートリクエスト"
      },
      "CritiqueOptimizeRequest": {
        "properties": {
          "api_key": {
            "type": "string",
            "title": "Api Key"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          },
          "book_id": {
            "type": "integer",
            "title": "Book Id"
          }
        },
        "type": "object",
        "required": [
          "api_key",
          "book_id"
        ],
        "title": "CritiqueOptimizeRequest",
        "description": "品質分析最適化リクエスト"
      },
      "EasyModeRequest": {
        "properties": {
          "api_key": {
            "type": "string",
            "title": "Api Key"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          },
          "genre": {
            "type": "string",
            "title": "Genre"
          },
          "keywords": {
            "type": "string",
            "title": "Keywords"
          },
          "archetype_key": {
            "type": "string",
            "title": "Archetype Key"
          },
          "target_eps": {
            "type": "integer",
            "title": "Target Eps"
          },
          "initial_limit": {
            "type": "integer",
            "title": "Initial Limit"
          },
          "word_count": {
            "type": "integer",
            "title": "Word Count"
          },
          "concept": {
            "type": "string",
            "title": "Concept",
            "default": ""
          },
          "tone_vibe": {
            "type": "number",
            "title": "Tone Vibe",
            "default": 0.6
          }
        },
        "type": "object",
        "required": [
          "api_key",
          "genre",
          "keywords",
          "archetype_key",
          "target_eps",
          "initial_limit",
          "word_count"
        ],
        "title": "EasyModeRequest",
        "description": "かんたんモード 全自動生成リクエスト"
      },
      "EpisodeGenerateCandidatesRequest": {
        "properties": {
          "api_key": {
            "type": "string",
            "title": "Api Key"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          },
          "book_id": {
            "type": "integer",
            "title": "Book Id"
          },
          "write_from": {
            "type": "integer",
            "title": "Write From"
          },
          "write_to": {
            "type": "integer",
            "title": "Write To"
          },
          "passion": {
            "type": "number",
            "title": "Passion"
          },
          "word_count": {
            "type": "integer",
            "title": "Word Count"
          },
          "do_refine": {
            "type": "boolean",
            "title": "Do Refine"
          },
          "env_state": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Env State"
          },
          "pipeline_mode": {
            "type": "boolean",
            "title": "Pipeline Mode",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "api_key",
          "book_id",
          "write_from",
          "write_to",
          "passion",
          "word_count",
          "do_refine"
        ],
        "title": "EpisodeGenerateCandidatesRequest",
        "description": "エピソード候補生成リクエスト"
      },
      "EpisodeGenerateRequest": {
        "properties": {
          "api_key": {
            "type": "string",
            "title": "Api Key"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          },
          "book_id": {
            "type": "integer",
            "title": "Book Id"
          },
          "write_from": {
            "type": "integer",
            "title": "Write From"
          },
          "write_to": {
            "type": "integer",
            "title": "Write To"
          },
          "passion": {
            "type": "number",
            "title": "Passion"
          },
          "word_count": {
            "type": "integer",
            "title": "Word Count"
          },
          "do_refine": {
            "type": "boolean",
            "title": "Do Refine"
          },
          "env_state": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Env State"
          },
          "pipeline_mode": {
            "type": "boolean",
            "title": "Pipeline Mode",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "api_key",
          "book_id",
          "write_from",
          "write_to",
          "passion",
          "word_count",
          "do_refine"
        ],
        "title": "EpisodeGenerateRequest",
        "description": "エピソード生成リクエスト"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlanGenerationRequest": {
        "properties": {
          "api_key": {
            "type": "string",
            "title": "Api Key"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          }
        },
        "type": "object",
        "required": [
          "api_key",
          "params"
        ],
        "title": "PlanGenerationRequest",
        "description": "企画生成リクエスト"
      },
      "PlotExpandCandidatesRequest": {
        "properties": {
          "api_key": {
            "type": "string",
            "title": "Api Key"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          },
          "book_id": {
            "type": "integer",
            "title": "Book Id"
          },
          "gen_from": {
            "type": "integer",
            "title": "Gen From"
          },
          "gen_to": {
            "type": "integer",
            "title": "Gen To"
          }
        },
        "type": "object",
        "required": [
          "api_key",
          "book_id",
          "gen_from",
          "gen_to"
        ],
        "title": "PlotExpandCandidatesRequest",
        "description": "プロット候補展開リクエスト"
      },
      "PlotExpandRequest": {
        "properties": {
          "api_key": {
            "type": "string",
            "title": "Api Key"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          },
          "book_id": {
            "type": "integer",
            "title": "Book Id"
          },
          "gen_from": {
            "type": "integer",
            "title": "Gen From"
          },
          "gen_to": {
            "type": "integer",
            "title": "Gen To"
          }
        },
        "type": "object",
        "required": [
          "api_key",
          "book_id",
          "gen_from",
          "gen_to"
        ],
        "title": "PlotExpandRequest",
        "description": "プロット展開リクエスト"
      },
      "PlotRebuildRequest": {
        "properties": {
          "api_key": {
            "type": "string",
            "title": "Api Key"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          }
        },
        "type": "object",
        "required": [
          "api_key",
          "params"
        ],
        "title": "PlotRebuildRequest",
        "description": "プロット再構築リクエスト"
      },
      "RefineEroticRequest": {
        "properties": {
          "api_key": {
            "type": "string",
            "title": "Api Key"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          },
          "book_id": {
            "type": "integer",
            "title": "Book Id"
          },
          "ep_num": {
            "type": "integer",
            "title": "Ep Num"
          },
          "intensity": {
            "type": "integer",
            "title": "Intensity",
            "default": 2
          },
          "platform_preset": {
            "type": "string",
            "title": "Platform Preset",
            "default": "kakuyomu_romance"
          }
        },
        "type": "object",
        "required": [
          "api_key",
          "book_id",
          "ep_num"
        ],
        "title": "RefineEroticRequest",
        "description": "官能研磨リクエスト"
      },
...
      "RetryFailedRequest": {
        "properties": {
          "api_key": {
            "type": "string",
            "title": "Api Key"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          },
          "book_id": {
            "type": "integer",
            "title": "Book Id"
          },
          "passion": {
            "type": "number",
            "title": "Passion"
          },
          "word_count": {
            "type": "integer",
            "title": "Word Count"
          }
        },
        "type": "object",
        "required": [
          "api_key",
          "book_id",
          "passion",
          "word_count"
        ],
        "title": "RetryFailedRequest",
        "description": "失敗エピソード再試行リクエスト"
      },
...
    }
  }
}